+++ example/android/app/src/main/kotlin/dev/dartoos/example/MainActivity.kt
@@ -1,0 +1,5 @@
+package dev.dartoos.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()